wip extracting hole and casing
